Editing report.md

+++ src/Main.java
@@ -1,8 +1,8 @@
 public class Main {
     public static void main(String[] args) {
-        long currentBalance = 2_000_000_000;
+        int currentBalance = 2_000_000_000;
         int remittance = 500_000_000;
-        long totalBalance = currentBalance + remittance;
+        int totalBalance = currentBalance + remittance;
         System.out.println(totalBalance);
     }
 }
